feat: optimize gslb server and vs creation

+++ f5_ctlr_agent/gtm/wideip.py
@@ -55,15 +55,28 @@ class GTMWideIP:
         """
         try:
             exist = self._gtm.wideips.a_s.a.exists(name=config['name'], partition=self._partition)
+            created_new = False
+            
             if not exist:
                 log.info('GTM: Creating wideip {}'.format(config['name']))
-                self._gtm.wideips.a_s.a.create(
-                    name=config['name'],
-                    partition=self._partition,
-                    lastResortPool="none",
-                    poolLbMode=config['LoadBalancingMode'])
-                self.attach_pool_to_wideip(config['name'], list(newPools.values()))
-            else:
+                try:
+                    self._gtm.wideips.a_s.a.create(
+                        name=config['name'],
+                        partition=self._partition,
+                        lastResortPool="none",
+                        poolLbMode=config['LoadBalancingMode'])
+                    self.attach_pool_to_wideip(config['name'], list(newPools.values()))
+                    created_new = True
+                except Exception as e:
+                    # Race condition: another thread created it between exists() and create()
+                    if 'already exists' in str(e).lower() or '409' in str(e):
+                        log.debug('GTM: WideIP {} already exists , updating it'.format(config['name']))
+                        exist = True  # Fall through to update path
+                    else:
+                        raise
+            
+            # Update existing wideIP (or handle race condition)
+            if exist and not created_new:
                 wideip = self._gtm.wideips.a_s.a.load(
                     name=config['name'],
                     partition=self._partition)
@@ -213,6 +226,17 @@ class GTMWideIP:
             # WideIP exists - proceed with delete
             try:
                 wideip = self._gtm.wideips.a_s.a.load(name=wideipName, partition=self._partition)
+                
+                # Fix lastResortPool if empty (BIG-IP API guard)
+                if wideip.lastResortPool == "":
+                    wideip.lastResortPool = "none"
+                    wideip.update()
+                
+                # Check if pools are still attached
+                if hasattr(wideip, 'pools') and len(wideip.pools) > 0:
+                    log.debug("GTM: Cannot delete wideIP {} - pools still attached".format(wideipName))
+                    return False
+                    
                 wideip.delete()
                 log.info("GTM: Deleted wideIP {}".format(wideipName))
                 return True
